Move conversion functions in separate file
